change points fields int->float64

+++ internal/entities/accounts.go
@@ -4,5 +4,5 @@ type Account struct {
 	ID       string
 	Login    string
 	HashPass string
-	Points   int
+	Points   float64
 }
